requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.1.3
et-xmlfile==1.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
openai==0.27.6
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests===2.28.2
six==1.16.0
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.7
Werkzeug==2.0.2
ibm-watson==7.0.0
langchain
streamlit
streamlit-chat-media
retry
duckduckgo-search
transformers
tabulate
wikipedia
faiss-cpu # if u dont have a GPU
nest_asyncio
torch
# tensorflow >= 2.0 -- For other future models
# flax -- For other future models
pytest-playwright
undetected-chromedriver>=3.2.1
markdownify>=0.11.6
hugchat
EdgeGPT
GoogleBard
lark
accelerate
diffusers
PyVirtualDisplay==3.0
revChatGPT==6.8.6
pinecone-client
unstructured==0.8.1
chromadb==0.3.29